+++ packages/fcl/src/default-config.js
@@ -1,7 +1,8 @@
 import {config} from "@onflow/sdk"
-import {SESSION_STORAGE} from "./config-utils"
+import {DISCOVERY_METHOD, STORAGE_DEFAULT, APP_DOMAIN_TAG} from "./config-utils"
 
 config({
-  "discovery.wallet.method.default": "IFRAME/RPC",
-  "fcl.storage.default": SESSION_STORAGE,
+  "discovery.wallet.method.default": DISCOVERY_METHOD,
+  "fcl.storage.default": STORAGE_DEFAULT,
+  "fcl.appDomainTag.default": APP_DOMAIN_TAG,
 })